Remove test symbols from the lib/events/athena package (#56490)
lib/events/athena/testhelpers.go caused testify to leak into production
binaries. Its entire contents has been moved into
lib/events/athena/integration_test.go since that was the only place
which consumed it.

Updates https://github.com/gravitational/teleport/issues/51023.
